+++ src/frontend/src/features/pip/components/controls/PipOptionsMenuItems.tsx
@@ -1,10 +1,4 @@
-import { Menu as RACMenu, MenuItem, MenuSection } from 'react-aria-components'
-import {
-  RiHand,
-  RiClosedCaptioningLine,
-  RiArrowUpLine,
-  RiEmotionLine,
-} from '@remixicon/react'
+import { Menu as RACMenu, MenuSection } from 'react-aria-components'
 import { useTranslation } from 'react-i18next'
 import { Separator } from '@/primitives/Separator'
 import { FeedbackMenuItem } from '@/features/rooms/livekit/components/controls/Options/FeedbackMenuItem'
@@ -12,12 +6,7 @@ import { EffectsMenuItem } from '@/features/rooms/livekit/components/controls/Op
 import { SupportMenuItem } from '@/features/rooms/livekit/components/controls/Options/SupportMenuItem'
 import { PictureInPictureMenuItem } from '@/features/rooms/livekit/components/controls/Options/PictureInPictureMenuItem'
 import { pipLayoutStore } from '@/features/pip/stores/pipLayoutStore'
-import { menuRecipe } from '@/primitives/menuRecipe'
-import { useRoomContext } from '@livekit/components-react'
-import { useRaisedHand } from '@/features/rooms/livekit/hooks/useRaisedHand'
-import { useSubtitles } from '@/features/subtitle/hooks/useSubtitles'
-import { useAreSubtitlesAvailable } from '@/features/subtitle/hooks/useAreSubtitlesAvailable'
-import { useSnapshot } from 'valtio'
+import { PipOverflowItems } from './PipOverflowItems'
 import type { CollapsibleControl } from '../PipControlBar'
 
 type PipOptionsMenuItemsProps = {
@@ -40,7 +29,7 @@ export const PipOptionsMenuItems = ({
       {hasOverflow && (
         <>
           <MenuSection>
-            <OverflowItems overflowControls={overflowControls} t={t} />
+            <PipOverflowItems overflowControls={overflowControls} t={t} />
           </MenuSection>
           <Separator />
         </>
@@ -55,62 +44,5 @@ export const PipOptionsMenuItems = ({
         <FeedbackMenuItem />
       </MenuSection>
     </RACMenu>
-  )
-}
-
-const OverflowItems = ({
-  overflowControls,
-  t,
-}: {
-  overflowControls: Set<CollapsibleControl>
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  t: any
-}) => {
-  const room = useRoomContext()
-  const { isHandRaised, toggleRaisedHand } = useRaisedHand({
-    participant: room.localParticipant,
-  })
-  const { areSubtitlesOpen, toggleSubtitles } = useSubtitles()
-  const areSubtitlesAvailable = useAreSubtitlesAvailable()
-  const pipSnap = useSnapshot(pipLayoutStore)
-  const toggleReactions = () => {
-    pipLayoutStore.showReactionsToolbar = !pipSnap.showReactionsToolbar
-  }
-  const itemClass = menuRecipe({ icon: true, variant: 'dark' }).item
-
-  return (
-    <>
-      {overflowControls.has('reactions') && (
-        <MenuItem onAction={toggleReactions} className={itemClass}>
-          <RiEmotionLine size={20} />
-          {t('controls.reactions.button')}
-        </MenuItem>
-      )}
-      {overflowControls.has('screenShare') && (
-        <MenuItem
-          onAction={() => {
-            /* screen share requires track toggle, handled externally */
-          }}
-          className={itemClass}
-        >
-          <RiArrowUpLine size={20} />
-          {t('controls.screenShare.start')}
-        </MenuItem>
-      )}
-      {overflowControls.has('subtitles') && areSubtitlesAvailable && (
-        <MenuItem onAction={toggleSubtitles} className={itemClass}>
-          <RiClosedCaptioningLine size={20} />
-          {areSubtitlesOpen
-            ? t('controls.subtitles.open')
-            : t('controls.subtitles.closed')}
-        </MenuItem>
-      )}
-      {overflowControls.has('hand') && (
-        <MenuItem onAction={toggleRaisedHand} className={itemClass}>
-          <RiHand size={20} />
-          {isHandRaised ? t('controls.hand.lower') : t('controls.hand.raise')}
-        </MenuItem>
-      )}
-    </>
   )
 }
